+++ lib/log-courier/client_tls.rb
@@ -38,17 +38,17 @@ module LogCourier
       @logger = @options[:logger]
 
       [:port, :ssl_ca].each do |k|
-        raise "[LogCourierClient] '#{k}' is required" if @options[k].nil?
+        fail "output/courier: '#{k}' is required" if @options[k].nil?
       end
 
-      raise '[LogCourierClient] \'addresses\' must contain at least one address' if @options[:addresses].empty?
+      fail 'output/courier: \'addresses\' must contain at least one address' if @options[:addresses].empty?
 
       c = 0
       [:ssl_certificate, :ssl_key].each do
         c += 1
       end
 
-      raise '[LogCourierClient] \'ssl_certificate\' and \'ssl_key\' must be specified together' if c == 1
+      fail 'output/courier: \'ssl_certificate\' and \'ssl_key\' must be specified together' if c == 1
     end
 
     def connect(io_control)
@@ -71,6 +71,7 @@ module LogCourier
       @recv_thread = Thread.new do
         run_recv io_control
       end
+      return
     end
 
     def disconnect
@@ -78,7 +79,16 @@ module LogCourier
       @send_thread.join
       @recv_thread.raise ShutdownSignal
       @recv_thread.join
+      return
     end
+
+    def send(signature, message)
+      # Add to send queue
+      @send_q << [signature, message.length].pack('A4N') + message
+      return
+    end
+
+    private
 
     def run_send(io_control)
       # Ask for something to send
@@ -108,29 +118,31 @@ module LogCourier
           @ssl_client.write message
         end
       end
+      return
     rescue OpenSSL::SSL::SSLError, IOError, Errno::ECONNRESET => e
-      @logger.warn("[LogCourierClient] SSL write error: #{e}") unless @logger.nil?
+      @logger.warn 'SSL write error', :error => e.message unless @logger.nil?
       io_control << ['F']
+      return
     rescue ShutdownSignal
-      # Just shutdown
-    rescue => e
-      @logger.warn("[LogCourierClient] Unknown SSL write error: #{e}") unless @logger.nil?
-      @logger.warn("[LogCourierClient] #{e.backtrace}: #{e.message} (#{e.class})") unless @logger.nil?
+      return
+    rescue StandardError, NativeException => e
+      @logger.warn e, :hint => 'Unknown SSL write error' unless @logger.nil?
       io_control << ['F']
+      return
     end
 
     def run_recv(io_control)
       loop do
         # Grab a header
         header = @ssl_client.read(8)
-        raise EOFError if header.nil?
+        fail EOFError if header.nil?
 
         # Decode signature and length
         signature, length = header.unpack('A4N')
 
         if length > 1048576
           # Too big raise error
-          @logger.warn("[LogCourierClient] Invalid message: data too big (#{length})") unless @logger.nil?
+          @logger.warn 'Invalid message: data too big', :data_length => length unless @logger.nil?
           io_control << ['F']
           break
         end
@@ -141,29 +153,28 @@ module LogCourier
         # Pass through to receive
         io_control << ['R', signature, message]
       end
+      return
     rescue OpenSSL::SSL::SSLError, IOError, Errno::ECONNRESET => e
-      @logger.warn("[LogCourierClient] SSL read error: #{e}") unless @logger.nil?
+      @logger.warn 'SSL read error', :error => e.message unless @logger.nil?
       io_control << ['F']
+      return
     rescue EOFError
-      @logger.warn("[LogCourierClient] Connection closed by server") unless @logger.nil?
+      @logger.warn 'Connection closed by server' unless @logger.nil?
       io_control << ['F']
+      return
     rescue ShutdownSignal
-      # Just shutdown
+      return
     rescue => e
-      @logger.warn("[LogCourierClient] Unknown SSL read error: #{e}") unless @logger.nil?
-      @logger.warn("[LogCourierClient] #{e.backtrace}: #{e.message} (#{e.class})") unless @logger.nil?
+      @logger.warn e, :hint => 'Unknown SSL read error' unless @logger.nil?
       io_control << ['F']
-    end
-
-    def send(signature, message)
-      # Add to send queue
-      @send_q << [signature, message.length].pack('A4N') + message
+      return
     end
 
     def pause_send
       return if @send_paused
       @send_paused = true
       @send_q << nil
+      return
     end
 
     def send_paused
@@ -175,44 +186,53 @@ module LogCourier
         @send_paused = false
         @send_q << nil
       end
+      return
     end
 
     def tls_connect
       # TODO: Implement random selection - and don't use separate :port - remember to update post_connection_check too
-      @logger.info("[LogCourierClient] Connecting to #{@options[:addresses][0]}:#{@options[:port]}") unless @logger.nil?
-      tcp_socket = TCPSocket.new(@options[:addresses][0], @options[:port])
+      address = @options[:addresses][0]
+      port = @options[:port]
 
-      ssl = OpenSSL::SSL::SSLContext.new
+      @logger.info 'Connecting', :address => address, :port => port unless @logger.nil?
 
-      unless @options[:ssl_certificate].nil?
-        ssl.cert = OpenSSL::X509::Certificate.new(File.read(@options[:ssl_certificate]))
-        ssl.key = OpenSSL::PKey::RSA.new(File.read(@options[:ssl_key]), @options[:ssl_key_passphrase])
+      begin
+        tcp_socket = TCPSocket.new(address, port)
+
+        ssl = OpenSSL::SSL::SSLContext.new
+
+        unless @options[:ssl_certificate].nil?
+          ssl.cert = OpenSSL::X509::Certificate.new(File.read(@options[:ssl_certificate]))
+          ssl.key = OpenSSL::PKey::RSA.new(File.read(@options[:ssl_key]), @options[:ssl_key_passphrase])
+        end
+
+        cert_store = OpenSSL::X509::Store.new
+        cert_store.add_file(@options[:ssl_ca])
+        #ssl.cert_store = cert_store
+        ssl.verify_mode = OpenSSL::SSL::VERIFY_PEER | OpenSSL::SSL::VERIFY_FAIL_IF_NO_PEER_CERT
+
+        @ssl_client = OpenSSL::SSL::SSLSocket.new(tcp_socket)
+
+        socket = @ssl_client.connect
+
+        # Verify certificate
+        socket.post_connection_check(address)
+
+        # Add extra logging data now we're connected
+        @logger['address'] = address
+        @logger['port'] = port
+
+        @logger.info 'Connected successfully' unless @logger.nil?
+        return
+      rescue OpenSSL::SSL::SSLError, IOError, Errno::ECONNRESET => e
+        @logger.warn 'Connection failed', :error => e.message, :address => address, :port => port unless @logger.nil?
+        return
+      rescue ShutdownSignal
+        return
+      rescue StandardError, NativeException => e
+        @logger.warn e, :hint => 'Unknown connection failure', :address => address, :port => port unless @logger.nil?
+        raise e
       end
-
-      cert_store = OpenSSL::X509::Store.new
-      cert_store.add_file(@options[:ssl_ca])
-      #ssl.cert_store = cert_store
-      ssl.verify_mode = OpenSSL::SSL::VERIFY_PEER | OpenSSL::SSL::VERIFY_FAIL_IF_NO_PEER_CERT
-
-      @ssl_client = OpenSSL::SSL::SSLSocket.new(tcp_socket)
-
-      socket = @ssl_client.connect
-
-      # Verify certificate
-      socket.post_connection_check(@options[:addresses][0])
-
-      @logger.info("[LogCourierClient] Connected successfully") unless @logger.nil?
-
-      socket
-    rescue OpenSSL::SSL::SSLError, IOError, Errno::ECONNRESET => e
-      @logger.warn("[LogCourierClient] Connection to #{@options[:addresses][0]}:#{@options[:port]} failed: #{e}") unless @logger.nil?
-    rescue ShutdownSignal
-      # Just shutdown
-      0
-    rescue StandardError, NativeException => e
-      @logger.warn("[LogCourierClient] Unknown connection failure to #{@options[:addresses][0]}:#{@options[:port]}: #{e}") unless @logger.nil?
-      @logger.warn("[LogCourierClient] #{e.backtrace}: #{e.message} (#{e.class})") unless @logger.nil?
-      raise e
     end
   end
 end