+++ lib/log-courier/event_queue.rb
@@ -33,7 +33,7 @@ module LogCourier
     # Creates a fixed-length queue with a maximum size of +max+.
     #
     def initialize(max)
-      raise ArgumentError, "queue size must be positive" unless max > 0
+      fail ArgumentError, "queue size must be positive" unless max > 0
       @max = max
       @enque_cond = ConditionVariable.new
       @num_enqueue_waiting = 0
@@ -44,20 +44,19 @@ module LogCourier
       self.taint
       @mutex = Mutex.new
       @cond = ConditionVariable.new
+      return
     end
 
     #
     # Returns the maximum size of the queue.
     #
-    def max
-      @max
-    end
+    attr_reader :max
 
     #
     # Sets the maximum size of the queue.
     #
     def max=(max)
-      raise ArgumentError, "queue size must be positive" unless max > 0
+      fail ArgumentError, "queue size must be positive" unless max > 0
 
       @mutex.synchronize do
         if max <= @max
@@ -90,7 +89,7 @@ module LogCourier
           ensure
             @num_enqueue_waiting -= 1
           end
-          raise TimeoutError if !timeout.nil? and Time.now - start >= timeout
+          fail TimeoutError if !timeout.nil? and Time.now - start >= timeout
         end
 
         @que.push obj
@@ -113,38 +112,13 @@ module LogCourier
     # Retrieves data from the queue and runs a waiting thread, if any.
     #
     def pop(*args)
-      retval = _pop_timeout *args
+      retval = pop_timeout *args
       @mutex.synchronize do
         if @que.length < @max
           @enque_cond.signal
         end
       end
       retval
-    end
-
-    #
-    # Retrieves data from the queue.  If the queue is empty, the calling thread is
-    # suspended until data is pushed onto the queue or, if set, +timeout+ seconds
-    # passes.  If +timeout+ is 0, the thread isn't suspended, and an exception is
-    # raised.
-    #
-    def _pop_timeout(timeout = nil)
-      unless timeout.nil?
-        start = Time.now
-      end
-      @mutex.synchronize do
-        loop do
-          return @que.shift unless @que.empty?
-          raise TimeoutError if timeout == 0
-          begin
-            @num_waiting += 1
-            @cond.wait @mutex, timeout
-          ensure
-            @num_waiting -= 1
-          end
-          raise TimeoutError if !timeout.nil? and Time.now - start >= timeout
-        end
-      end
     end
 
     #
@@ -189,6 +163,34 @@ module LogCourier
     #
     def num_waiting
       @num_waiting + @num_enqueue_waiting
+    end
+
+    private
+
+    #
+    # Retrieves data from the queue.  If the queue is empty, the calling thread is
+    # suspended until data is pushed onto the queue or, if set, +timeout+ seconds
+    # passes.  If +timeout+ is 0, the thread isn't suspended, and an exception is
+    # raised.
+    #
+    def pop_timeout(timeout = nil)
+      unless timeout.nil?
+        start = Time.now
+      end
+      @mutex.synchronize do
+        loop do
+          return @que.shift unless @que.empty?
+          fail TimeoutError if timeout == 0
+          begin
+            @num_waiting += 1
+            @cond.wait @mutex, timeout
+          ensure
+            @num_waiting -= 1
+          end
+          fail TimeoutError if !timeout.nil? and Time.now - start >= timeout
+        end
+      end
+      return
     end
   end
 end