+++ lib/log-courier/server_tcp.rb
@@ -24,12 +24,24 @@ require 'thread'
 module LogCourier
   # Wrap around TCPServer to grab last error for use in reporting which peer had an error
   class ExtendedTCPServer < TCPServer
-    # Yield the peer
+    attr_reader :peer
+
+    def initialise
+      reset_peer
+      super
+    end
+
+    # Save the peer
     def accept
       sock = super
       peer = sock.peeraddr(:numeric)
-      Thread.current['LogCourierPeer'] = "#{peer[2]}:#{peer[1]}"
+      @peer = "#{peer[2]}:#{peer[1]}"
       return sock
+    end
+
+    def reset_peer
+      @peer = 'unknown'
+      return
     end
   end
 
@@ -57,11 +69,11 @@ module LogCourier
 
       if @options[:transport] == 'tls'
         [:ssl_certificate, :ssl_key].each do |k|
-          raise "[LogCourierServer] '#{k}' is required" if @options[k].nil?
+          fail "input/courier: '#{k}' is required" if @options[k].nil?
         end
 
         if @options[:ssl_verify] and (!@options[:ssl_verify_default_ca] && @options[:ssl_verify_ca].nil?)
-          raise '[LogCourierServer] Either \'ssl_verify_default_ca\' or \'ssl_verify_ca\' must be specified when ssl_verify is true'
+          fail 'input/courier: Either \'ssl_verify_default_ca\' or \'ssl_verify_ca\' must be specified when ssl_verify is true'
         end
       end
 
@@ -94,16 +106,18 @@ module LogCourier
             ssl.verify_mode = OpenSSL::SSL::VERIFY_PEER | OpenSSL::SSL::VERIFY_FAIL_IF_NO_PEER_CERT
           end
 
+          # Create the OpenSSL server - set start_immediately to false so we can multithread handshake
           @server = OpenSSL::SSL::SSLServer.new(@tcp_server, ssl)
+          @server.start_immediately = false
         else
           @server = @tcp_server
         end
 
         if @options[:port] == 0
-          @logger.warn '[LogCourierServer] Transport ' + @options[:transport] + ' is listening on ephemeral port ' + @port.to_s unless @logger.nil?
+          @logger.warn 'Ephemeral port allocated', :transport => @options[:transport], :port => @port unless @logger.nil?
         end
       rescue => e
-        raise "[LogCourierServer] Failed to initialise: #{e}"
+        raise "input/courier: Failed to initialise: #{e}"
       end
     end # def initialize
 
@@ -111,20 +125,20 @@ module LogCourier
       client_threads = {}
 
       loop do
-        # This means ssl accepting is single-threaded.
+        # Because start_immediately is false, TCP accept is single thread but
+        # handshake is essentiall multithreaded as we defer it to the thread
+        @tcp_server.reset_peer
+        client = nil
         begin
           client = @server.accept
         rescue EOFError, OpenSSL::SSL::SSLError, IOError => e
-          # Handshake failure or other issue
-          peer = Thread.current['LogCourierPeer'] || 'unknown'
-          @logger.warn "[LogCourierServer] Connection from #{peer} failed to initialise: #{e}" unless @logger.nil?
-          client.close rescue nil
+          # Accept failure or other issue
+          @logger.warn 'Connection failed to accept', :error => e.message, :peer => @tcp_server.peer unless @logger.nil
+          client.close rescue nil unless client.nil?
           next
         end
 
-        peer = Thread.current['LogCourierPeer'] || 'unknown'
-
-    	  @logger.info "[LogCourierServer] New connection from #{peer}" unless @logger.nil?
+    	  @logger.info 'New connection', :peer => @tcp_server.peer unless @logger.nil?
 
         # Clear up finished threads
         client_threads.delete_if do |_, thr|
@@ -132,17 +146,16 @@ module LogCourier
         end
 
         # Start a new connection thread
-        client_threads[client] = Thread.new(client, peer) do |client_copy, peer_copy|
-          ConnectionTcp.new(@logger, client_copy, peer_copy, @options).run(&block)
+        client_threads[client] = Thread.new(client, @tcp_server.peer) do |client_copy, peer_copy|
+          run_thread client_copy, peer_copy, &block
         end
       end
+      return
     rescue ShutdownSignal
-      # Capture shutting down signal
-      0
+      return
     rescue StandardError, NativeException => e
       # Some other unknown problem
-      @logger.warn("[LogCourierServer] Unknown error: #{e}") unless @logger.nil?
-      @logger.warn("[LogCourierServer] #{e.backtrace}: #{e.message} (#{e.class})") unless @logger.nil?
+      @logger.warn e, :hint => 'Unknown error, shutting down' unless @logger.nil?
       raise e
     ensure
       # Raise shutdown in all client threads and join then
@@ -153,6 +166,24 @@ module LogCourier
       client_threads.each(&:join)
 
       @tcp_server.close
+    end
+
+    private
+
+    def run_thread(client, peer, &block)
+      # Perform the handshake inside the new thread so we don't block TCP accept
+      if @options[:transport] == 'tls'
+        begin
+          client.accept
+        rescue EOFError, OpenSSL::SSL::SSLError, IOError => e
+          # Handshake failure or other issue
+          @logger.warn 'Connection failed to initialise', :error => e.message, :peer => peer unless @logger.nil?
+          client.close
+          return
+        end
+      end
+
+      ConnectionTcp.new(@logger, client, peer, @options).run(&block)
     end
   end
 
@@ -179,7 +210,7 @@ module LogCourier
 
         # Sanity
         if length > @options[:max_packet_size]
-          raise ProtocolError, "packet too large (#{length} > #{@options[:max_packet_size]})"
+          fail ProtocolError, "packet too large (#{length} > #{@options[:max_packet_size]})"
         end
 
         # While we're processing, EOF is bad as it may occur during send
@@ -198,31 +229,60 @@ module LogCourier
         # If we EOF next it's a graceful close
         @in_progress = false
       end
+      return
     rescue TimeoutError
       # Timeout of the connection, we were idle too long without a ping/pong
-      @logger.warn("[LogCourierServer] Connection from #{@peer} timed out") unless @logger.nil?
+      @logger.warn 'Connection timed out', :peer => @peer unless @logger.nil?
+      return
     rescue EOFError
       if @in_progress
-        @logger.warn("[LogCourierServer] Premature connection close on connection from #{@peer}") unless @logger.nil?
+        @logger.warn 'Unexpected EOF', :peer => @peer unless @logger.nil?
       else
-        @logger.info("[LogCourierServer] Connection from #{@peer} closed") unless @logger.nil?
+        @logger.info 'Connection closed', :peer => @peer unless @logger.nil?
       end
+      return
     rescue OpenSSL::SSL::SSLError, IOError, Errno::ECONNRESET => e
       # Read errors, only action is to shutdown which we'll do in ensure
-      @logger.warn("[LogCourierServer] SSL error on connection from #{@peer}: #{e}") unless @logger.nil?
+      @logger.warn 'SSL error, connection aborted', :error => e.message, :peer => @peer unless @logger.nil?
+      return
     rescue ProtocolError => e
       # Connection abort request due to a protocol error
-      @logger.warn("[LogCourierServer] Protocol error on connection from #{@peer}: #{e}") unless @logger.nil?
+      @logger.warn 'Protocol error, connection aborted', :error => e.message, :peer => @peer unless @logger.nil?
+      return
     rescue ShutdownSignal
       # Shutting down
-      @logger.warn("[LogCourierServer] Closing connecting from #{@peer}: server shutting down") unless @logger.nil?
-    rescue => e
+      @logger.info 'Server shutting down, closing connection', :peer => @peer unless @logger.nil?
+      return
+    rescue StandardError, NativeException => e
       # Some other unknown problem
-      @logger.warn("[LogCourierServer] Unknown error on connection from #{@peer}: #{e}") unless @logger.nil?
-      @logger.warn("[LogCourierServer] #{e.backtrace}: #{e.message} (#{e.class})") unless @logger.nil?
+      @logger.warn e, :hint => 'Unknown error, connection aborted', :peer => @peer unless @logger.nil?
+      return
     ensure
       @fd.close rescue nil
     end
+
+    def send(signature, message)
+      reset_timeout
+      data = signature + [message.length].pack('N') + message
+      done = 0
+      loop do
+        begin
+          written = @fd.write_nonblock(data[done...data.length])
+        rescue IO::WaitReadable
+          fail TimeoutError if IO.select([@fd], nil, [@fd], @timeout - Time.now.to_i).nil?
+          retry
+        rescue IO::WaitWritable
+          fail TimeoutError if IO.select(nil, [@fd], [@fd], @timeout - Time.now.to_i).nil?
+          retry
+        end
+        fail ProtocolError, "write failure (#{done}/#{data.length})" if written == 0
+        done += written
+        break if done >= data.length
+      end
+      return
+    end
+
+    private
 
     def recv(need)
       reset_timeout
@@ -231,16 +291,16 @@ module LogCourier
         begin
        	  buffer = @fd.read_nonblock need - have.length
         rescue IO::WaitReadable
-          raise TimeoutError if IO.select([@fd], nil, [@fd], @timeout - Time.now.to_i).nil?
+          fail TimeoutError if IO.select([@fd], nil, [@fd], @timeout - Time.now.to_i).nil?
           retry
         rescue IO::WaitWritable
-          raise TimeoutError if IO.select(nil, [@fd], [@fd], @timeout - Time.now.to_i).nil?
+          fail TimeoutError if IO.select(nil, [@fd], [@fd], @timeout - Time.now.to_i).nil?
           retry
         end
         if buffer.nil?
-          raise EOFError
+          fail EOFError
         elsif buffer.length == 0
-          raise ProtocolError, "read failure (#{have.length}/#{need})"
+          fail ProtocolError, "read failure (#{have.length}/#{need})"
         end
         if have.length == 0
           have = buffer
@@ -252,29 +312,10 @@ module LogCourier
       have
     end
 
-    def send(signature, message)
-      reset_timeout
-      data = signature + [message.length].pack('N') + message
-      done = 0
-      loop do
-        begin
-          written = @fd.write_nonblock(data[done...data.length])
-        rescue IO::WaitReadable
-          raise TimeoutError if IO.select([@fd], nil, [@fd], @timeout - Time.now.to_i).nil?
-          retry
-        rescue IO::WaitWritable
-          raise TimeoutError if IO.select(nil, [@fd], [@fd], @timeout - Time.now.to_i).nil?
-          retry
-        end
-        raise ProtocolError, "write failure (#{done}/#{data.length})" if written == 0
-        done += written
-        break if done >= data.length
-      end
-    end
-
     def reset_timeout
       # TODO: Make configurable
       @timeout = Time.now.to_i + 1_800
+      return
     end
   end
 end