+++ lib/logstash/inputs/courier.rb
@@ -59,6 +59,9 @@ module LogStash
       # Max packet size
       config :max_packet_size, :validate => :number
 
+      # Peer recv queue
+      config :peer_recv_queue, :validate => :number
+
       public
 
       def register
@@ -75,10 +78,11 @@ module LogStash
           ssl_verify:            @ssl_verify,
           ssl_verify_default_ca: @ssl_verify_default_ca,
           ssl_verify_ca:         @ssl_verify_ca,
-          curve_secret_key:      @curve_secret_key
+          curve_secret_key:      @curve_secret_key,
         }
 
         options[:max_packet_size] = @max_packet_size unless @max_packet_size.nil?
+        options[:peer_recv_queue] = @peer_recv_queue unless @peer_recv_queue.nil?
 
         require 'log-courier/server'
         @log_courier = LogCourier::Server.new options